Update MscmpSystForms typespecs and documentation
This commit goes a long way to completing the documentation for
MscmpSystForms.  We include both typespecs and documentation in the
commit as for all practical purposes the typespecs are more a form of
documentation for us than any other purpose they might serve.

Left undocumented in MscmpSystForms are the
`MscmpSystForms.WebComponent` components.  There's a token effort to put
something down, but nothing that really counts as "documentation".

There will also be a couple of code changes, but primarily related to
naming in a place or two.