Update Msplatform, MssubMcp, & MssubBms docs
This update is mostly focused on ensuring that configurations are
correct for creating the documentation for the subsystem and platform
levels of the application.  We re-generate what exists of this
documentation, but don't stake the claim that this documentation is
complete in any sense of the word.

The expectation is there's a fair amount of work to do here.