Disable MscmpSystForms/Credo LongQuoteBlocks check
The `__using__` macro of `MscmpSystForms` is pretty long because we
include documentation for the functions we create in using modules.
This results in a Credo warning about doing too much in `quote` blocks.

Credo's primary concern is correct: do too much in a quoted block and
it may mean you're doing too much in the macro that should be in
functions elsewhere.  But that's not the problem in our case, we're
actually doing fairly little and even that is mostly just delegation.
Credo does allow ignoring comments in the check, but these are @doc
attributes, not Elixir comments.

In this case, we'll one-off ignore the Credo check since it really
doesn't apply to what we're doing.  We will however not configure Credo
differently on the whole because the check, and its constraints, in the
context it's targeted at is valid.