Fix Credo alias depth complaint
Credo is complaining about a reference to `Phoenix.HTML.Forms.to_form/2`
as being something to consider aliasing.  In this case, we only use it
once and I think being explicit here is more helpful than being brief.

Let's try setting the configuration for this Credo check overall to only
fire a warning if there are more than 3 instances of the same aliasing
opportunity; 4 usages seems a good point at which to consider requiring
an alias.

Change across the project.