Fix minor "on-premise" usage issue
The phrase "on-premise" isn't correct for in that it uses the word
"premise" incorrectly.  Premise has no connection to meanings of
"location"; "premises" does have a meaning of location.

To be fair, the technology industry by-and-large misuses the term
"premise" in this way and for the purpose we were doing so.  I expect
that at some point this usage will make its way into the dictionary as
well... but that hasn't happened just yet.

Therefore, for now we'll be pedants and use the more correct
"on-premises".