Fix Msplatform production config host/port
The Phoenix default port for production is 4000, which is already
assigned by the IANA.  Our port range is 4380-4388, which is unassigned.

For production MCP, we'll designate a default port of 4380.  This should
be set in prod.exs, but more importantly in runtime.exs.

Also, remove "example.com" from prod.exs and add "localhost".  Really
we have instructions to configure runtime.exs in the deployment guide
and rely on `check_origin`.