Fix GitHub actions path to bootstrap_database.psql
We've removed this script from each Component and made it a single
script under tools.  Our GitHub actions config however wasn't updated
when this change was made and so tests are failing.

Hopefully what we've done is correct... or there will be other commits
to get this working on GitHub.