Update project dependencies (across all projects)
This is response to a GitHub Actions CI failure in Postgrex.  Postgrex
had a problem which was causing the MscmpSystDb dialyzer run to fail.
This had been fixed in a later patch update to postgrex which was not
the version we were locked to.

Let's go ahead and do a bulk update.  We'll do it sooner or later anyway
and if we're going to do a first testing release with this code, we
should test with the latest version of the dependencies in intend to run
with.