Fix remaining Netlify documentation issues
We've resolved the build issues with the last commit.  But Netlify
really doesn't want us routing to a sub-directory; in truth I think we'd
just need to set up a _redirects file (or something like that) somewhere
and we'd accomplish our goal... but for now we'll punt on it since we
don't have any more documentation right now anyway.