Finish Theory Book business systems defined
At least an initial pass at getting something here.  I'm not too happy
with the "Inferences" section and the whole approach here may need to be
rethought.

However, what's committed should be good enough to at least capture the
basic gist that there are three areas of concern, which occasionally
conflict in their nature, and which can lead to conflicts when the
boundaries aren't clear.