Refactor MscmpSystAuthn; ext. structs Credo config
Note that this commit is effectively part one of the work we need to do
in MscmpSystAuthn.  Here we deal with dependencies which have
implemented structs in place of maps (MscmpSystDb really); but there are
a number of map based data structures defined by MscmpSystAuthn that
should also be represented by structs and are not in this commit.  Here
we focus on:

  - Update the testing and dev_support references to MscmpSystDb data
    from maps to structs.

  - Set Credo strict module layout config.  Satisfy any Credo
    complaints.

Finally, we should also note that this commit passes testing, dialyzer,
and credo checks independent of the changes still needing to be
implemented.