Add MscmpSystAuthn disallowed passwords redaction
On thinking about this a bit, it's clear that one could use queries to
the disallowed passwords table to reveal passwords.  This commit aims to
help reduce the exposure that might happen related to logging of
disallowed password related queries.

The passwords and queries are all based on SHA-1 hashes of the passwords
which are disallowed.  Checks of actual user passwords in this table
causes us to take the user provided password, SHA-1 hash it, and then
check the table.  The SHA-1 hash of the user password is then checked
against the disallowed passwords table.  The issue is that if we log the
query, then then user password will be logged as a unsalted SHA-1 hash
which is substantially less secure than our authentication hashes which
are both salted and using a secure hashing algorithm.

This commit doesn't fully solve the problem, but is certainly a simple
step that could help reduce it.