Fix MssubMcp types; bootstrap & session name
In practice, many of these elements are only eventually required and
the mfa_credential may be `nil`.

Also, re-expose the MscmpSystSession `session_name` type.  Arguably we
should expect that MscmpSystSession will be a dependency of the
Msplatform apps, but we'll just do this for now to avoid adding another
dependency to the user interface layer (regardless how integral).