Add initial Theory Book BRM docs
We're far from having this flushed out, but this is a start at least.  I
think with some slightly better organization we'll be in the position
we're currently aiming for: getting the ideas out of my head and written
somewhere... well written and easy to read somewhere is still a goal for
a different day.