Reorg Theory Book BRM section
Try to flatten the structure.

We're not trying to justify having a whole book, but trying to convey
our important ideas as tersely as possible while still getting the
essentials across.  While the organizer in me prefers the more
structured approach we're discarding here, I think the flatter structure
is easier to read and get an overview when reading from the summary
pages.

While we're at it, refine some of the wording.