Update docs landing page books list layout
We're not publishing all of the books in the documentation out of the
gate, so it is helpful if the landing page lays out the published books
both dynamically and attractively.

Adding a span with some flex layout seems to do the trick; though this
is only working with statically added books unlike the menu which drives
off the front pages of each book being published or not.  Something to
think about