Update bootstrap_database.psql and GitHub Actions
While CI testing recent documentation changes, we ended up getting a
rash of DB permission related failures.  I've not completely gotten to
the root as to why this just started happening (the documentation
changes did not include any database or CI configuration changes).

This commit hopes to fix and improve the bootstrap_database.psql script
in the following ways.

  - In the GitHub actions configuration, don't use the
    `ms_syst_privileged` database role as the CI testing superuser.
    Have CI create a special DBA role for the purpose.  This mirrors
    the flow of a random DBA running the bootstrap_database.psql script;
    the script is suppose to create `ms_syst_privileged` itself.

  - Update bootstrap_database.psql to require that the password for
    `ms_syst_privileged` be set using a psql variable
    (`dbadmin_password`).

  - Update the GitHub actions configuration bootstap_database.psql calls
    to use both the new testing superuser for login and to set the psql
    variable discussed above.