Fix (attempt) for CI database GRANT failures
One change since our last successful PR is that the "latest" PostgreSQL
image pulled by GitHub Actions is now PostgreSQL 16 instead of 15.

PostgreSQL 16 changes the way roles can grant other roles.  I'm not sure
why this is an issue given that we gave the WITH ADMIN OPTION to the
`ms_syst_privilege` role which is trying to grant `ms_syst_managable` to
other roles... but it seems to be.

So lets see if we can clear this up by having CI test using PostgreSQL
15 instead.