Fix (again) PostgreSQL 16 ADMIN OPTION bug
The last CI run using PostgreSQL 15 worked, this is a PostgreSQL 16
breaking change for our scripts.

This commit lets PostgreSQL 16 load again (we'll have to bite this
bullet someday, may as well be today).  We did previously grant the
wrong role `ADMIN OPTION`; previously we granted `ms_syst_privileged`,
but that role is calling a `SECURITY DEFINER` function to create the
roles. The role that should have been granted ADMIN OPTION is
`ms_syst_owner`.

  - Change the GitHub Actions config to go for postgres latest (16 now)
    again since we understand this problem has to do with changes in the
    latest PostgreSQL version.

  - Switch the `ms_syst_managable` ADMIN OPTION to be granted to
    `ms_syst_owner`.

  - Add ADMIN OPTION to `ms_syst_access` for `ms_syst_owner`.  I expect
    it to ultimately have the same granting delegation issue as
    `ms_syst_managable`.