Fix (attempt) for MssubMcp stop failure
Having fixed the PostgreSQL 16 ADMIN OPTION breaking change, MssubMcp
still fails CI testing on `MscmpSyst.stop_all_applications/0`.  We're
getting a GenServer exit that we're not expecting (at least the way
we're getting it).  This issue doesn't replicate in testing locally.

CI is running Elixir 1.15.6, but locally we're at 1.15.5.  Let's see if
we have another regression issue.  This commit doesn't test OTP well...
I could have a regressing with the OTP version being newer in CI, but we
can start with Elixir.