Add temporary documentation for trigger naming
I am always looking for the notes on trigger naming in the long list of
trigger related commits.  We have a place for documentation so at least
get it in the documentation... even if not done with any elegance.

Later we'll want to clean this up, but for now this is better than where
it has been.