Fix reset_dev_database.psql for PostgreSQL 16
This is related to the PostgreSQL 16 change about admin role delegation.

Previously we fixed this same issue in the bootstrap database script,
but we never addressed the database reset scripts; these are just
helpers used during development and they don't get tested in CI or
otherwise.  We find problems with the reset script during active
development and the Pg16 issue cropped up this time.

Update all scripts where we reset the database.  They are all very
similar to each other so do this is a bulk update/commit.