Add skeleton MscmpSystNetwork component to project
This component is a fork of the net_address IP/MAC address utility
library by Isaac Yonemoto:

https://github.com/ityonemo/net_address

We were originally depending on that library directory, but with more
recent versions of Elixir we're getting compile time warnings.  I
submitted a PR to resolve the most concerning warning, but after a month
I've received no response so I'm assuming that the library is no longer
maintained.

The license of the library is a permissive license which should be
compatible with our current project license (or any license we're
currently contemplating).

We'll take most of the code as the library works well, but we'll make
that code consistent with the rest of our project.

This commit