Add MscmpSystNetwork logic and API
At this point the code is complete and represents the functionality that
we use in the application with maybe a couple other functions we might
not use (but were simple enough to put in).  Really we just want the
ability to represent IP addresses a bit nicer than with just the Erlang
`inet` library capabilities and the string parsing/sigil will make
working with IP addresses in places like tests nicer.

This commit includes documentation markdown, but not the actual
generated documentation.

We have doctests but no other tests.  That's what we'll get in place
next.