Fix MscmpSystNetwork/Cidr parse bug
The code was too forgiving of malformed input, allowing something like
"10.1.1.10/32/24" to be submitted successfully.  In that case we cannot
tell which network mask length is the desired one and assuming the first
will lead to the worst sort of bug: things working, but working
incorrectly or outside assumptions.

Just make such a poorly formed string blow-up.