Add more MscmpSystNetwork guards and use them
In the process of working on the MscmpSystAuthn refactor to
MscmpSystNetwork, it became clear that more guards would be useful.  So
create those guards for both IPv4 & IPv6.

Update the MscmpSystDb functionality related to MscmpSystNetwork to make
use of the new guards.

Also, add some `moduledoc`'s for the MscmpSystNetwork.Guards module.

Finally, change the private guards to be `defguardp`.