Fix mask bug in MscmpSystNetwork to_struct/2 calls
At some point we switched between allowing the `mask` attribute of the
IP address structs to be nil vs. now requiring they have some discrete
and valid value.  The to_struct functions were written when they were
assumed implicitly the host value when nil and so didn't bother to set
the mask explicitly.

This commit makes setting the mask explicit in the returned value.  We
will still infer that the mask is the single host mask value as the
callers of this function will most commonly not have an explicit mask
themselves.

Fix in both the IpV4 and IpV6 versions of the function.