Refactor Elixir Components for new project struct
This commit implements a change in the organization of our Elixir
Component applications.  Before this change we kept "API" files, files
containing code which defined the API and data that users external to
the Component would use to interface with it, in the `lib` folder
directly and then implementations of logic in the `impl` or `runtime`
directories.  As the projects became more complex or diverse in the ways
that you might interface with them from external code, having multiple
API files and even directories of API defining code was becoming less
intuitive.  This meant that a developer had to be aware of more and more
exceptions to what was considered API and what was considered a home for
implementation logic.

This commit reorganizes the code to more explicitly express what is
considered public API and what is related to implementation and runtime
logic.

  - Add a new common directory `lib/api` which serves as the home for
    all public API defining codes.  There may be additional structure
    inside the `api` directory as needed, but no implementation logic
    should be anywhere within `api` unless complying with this rule is
    unreasonable.  The spirit of this directory is that if something
    changes in it, there's a good chance you've got a semver version
    bump.

  - Reorganize all Components, Subsystems, and Platform code to comply
    with the new structure.  The exception to this is the Platform
    msapp_mcp_web app which conforms more closely to the standard
    Phoenix project layout than our project layout.

  - Split msdata and msform directories into their api and
    implementation focuses and place them into `api` or `impl` as
    appropriate.

Along the way some other tasks were addresses.

  - In older components such as MscmpSystSettings and MscmpSystEnums
    Ecto Schema definitions were co-located with their validation logic.
    While this is a Elixir standard practice it doesn't match our more
    strict separation between API definitions and implementation logic.

  - Split msforms along API/implementation lines.

There are additional smaller changes related to consistency, error
correction and formatting.

At this commit Elixir projects in the MuseBMS are passing tests, Credo
checks, and Dialyzer checks when run locally.