Fix attempt MscmpSystError testing
Apparently `runner.os` isn't available at this point in the CI workflow.

We aren't building on anything else than latest Ubuntu, so lets just
remove that until we get things reasonably working.  If we do need that
it looks like we could set up a matrix for os and then access the
`matrix.os` value.