Fix attempt 3 MscmpSystError CI testing
Kinda grasping at straws.  It's saying mix dialyzer isn't found now, but
it should be there (other mix related commands are working fine).  I can
only think that the working directory or that multiple commands in a run
are somehow not really respecting the working directory properly... but
there doesn't appear to be sufficient information as to why it wouldn't
work.