Fix attempt 7 GitHub Actions workflow
Now GitHub is telling me there is a syntax error in my workflow file
where there appears to be none.  I assume this means syntax error in
the 1000+ line file that sits under the "jobs" key which is the line
which says it has an error.

So, we had one job working try to add job by job until it fails.