Add explicit PLT paths to mix.exs Dialyzer config
It was looking like PLTs were being cached and not used suggesting that
the path did need to be explicitly set up like it was in MscmpSystError.

Add that configuration as well as update .gitignore to ignore the PLT
directory since the way we've done this will also cause our dialyzer
runs to look for priv/plts.