Remove underspecs from Dialyzer configs
We're getting errors in CI; some of this is because we added check
options to the mix.exs Dialyzer configuration.  However, some of the
now reported errors appear to be legitimately spurious: saying a
function doesn't return `nil` when it actually does in some
circumstances isn't helpful.  Normally, I'd want to disable the check
for that function, but adding :no_underspecs for that function seems to
be ignored whereas removing it globally removes the error.

Ideally I'd like to have underspecs on, but what it's protecting against
right now isn't enough to make it worth the trouble.