Fix attempt Dialyzer error MscmpSystSettings
Still getting spurious Dialyzer errors during CI:

"Warning: The expression produces multiple types, but none are matched."

This warning is not produced running Dialyzer locally with the same
OTP/Elixir versions as CI.   Unfortunately, that error is all I'm
getting; no indication were the error is appearing or other information
which might point to where in the code it's triggering.

This commit is an attempt to get better details from the CI Dialyzer
run.  Not sure if that will happen, but its worth a shot.