Refactor MscmpCoreGroups to MscmpCoreHierarchy
This is a larger change than just a renaming of a Component.  The
original MscmpCoreGroups contained two concepts: 1) a way to define a
hierarchical group structure and validate that its requirements are
being met; and 2) the definition of the actual groups, which must comply
with the configured standard to be considered valid.  The idea from here
was to have anything that looked like a group (menus, product
categories, chart of account structures, etc.) to be stored altogether
and validated against the common structure with only some specialized
concessions to handling "leaf nodes" (the linkages to the specialized
data to the general groups data.

On reflection, it became apparent that the hierarchy configuration piece
was large enough and more generally useful if it was independent from
the group concept.  Also, the generalized group concept, while workable,
may cause too many concessions on what is acceptable or not because of
the generalization.  Given this each specialized grouping need should be
addressed in a more specialized way allowing for validation within the
generalized hierarchical validation framework.

The actions here correct the issues listed above.  This commit includes:

  - Split the generalized hierarchy configuration concept into its own
    Component module: MscmpCoreHierarchy

  - Remove the MscmpCoreGroups Component.  This was forcing
    generalization at a place where specialization needs to rule the
    day.  Remove any associated documentation.

  - Complete the MscmpCoreHierarchy Component database structure and add
    some test data to be sure things are at least reasonably working at
    this juncture.

Note that this commit does a lot for the database side of the Component
and very little for the Elixir/app_server side.  We have just enough
Elixir code to start up the test suite, migrate the database, and start
the enum service (which is a dependency).  Nothing else exists in this
commit.