Refactor Hierarchy allow_node_refs naming
A Hierarchy implementing Component will ultimately be associating "item"
information with the grouping records which conform to the Hierarchy.
We've not really had a terribly well defined concept of what these
"items" are in our model.  To this point we've been just been using the
term "node" to call these something, but avoided doing so in any formal
or integrated conceptual way.

With this commit we're introducing the concept of "Branch Nodes" and
"Leaf Nodes".  Branch Nodes are the records defining grouping structures
and are the records the Hierarchy Component is most concerned.  Leaf
Nodes are the "item" level records which the Hierarchy/Branch Nodes are
organizing.

With this conceptual introduction, it seems right to give the
`ms_appl.data.conf_hierarchy_items.allow_node_refs` column a better name
which reflects the idea at stake better.  The new name for the column is
`allow_leaf_nodes`.