Refactor MscmpSystError testing for test standards
This is to introduce the better defined testing standard recently
developed.

The most substantial change, which is related to the new test standard,
is the division between unit tests and integration tests.  Unit tests
are focused on asynchronous, individual testing of the private API
functions of `impl` or `runtime` name-spaced modules; this usually means
that any database related operations have a richly seeded set of test
data loaded prior to running tests.  Integration Tests on the other hand
are focused on the synchronous testing of the Component's public API
functions ordered to be representative of the Component's supported
business process; integration tests seed only a bare minimum of test
data in the database with the expectation that the integration testing
itself will build much of the test data in earlier tests for use in
later tests.

To this end, this commit does the following:

  - Introduce already implemented parts of this standard that developed
    subsequent to the writing of the MscmpSystError Component.  This is
    mostly reflected in `test_helper.exs`.

  - Rename existing unit tests to be consistent with later naming
    standards.  For this Component that's mainly file/module naming.

  - Add `doctests_test.exs`

  - Add `integration_test.exs`