Refactor database reset scripts across project
Each individual database accessing component has a set of common reset
tasks that must be performed to clean up the database after testing or
using the DevSupport database for interactive development sessions; but
there's also a lot of common stuff.  Up to now the common stuff was cut
and paste into each Component's reset script.

This commit allows the Component level reset script to call a common
reset script which contains the common stuff.  Naturally this reduces
duplication and the related maintenance burden when the common stuff
needs to change.

Also, add DevSupport cleanup tasks to the common database reset script.
Previously, if DevSupport was used and not properly stopped, you'd end
up with a lot of cruft in the database and if you tried to remove this
manually and did so incorrectly, you could find yourself in a spot where
the automated setup and cleanup processes couldn't work.