Add MscmpCoreHierarchy Impl.Hierarchy module
This is the data manipulation module for ConfHierarchy records, though
Hierarchy Types and Hierarchy States are included as they're support
data records.

Note that this is an initial set of functions and not complete for a
fully user configurable module.  That completion will likely be deferred
for some time since with just the Hierarchy related database in place we
probably have enough for our menu related Component (our real focus).

Also, add related unit tests.