Refactor MscmpCoreHierarchy to MscmpSystHierarchy
Previously we called this a core application module.  This made sense
when the generic concept of "group" included the hierarchy configuration
and validation framework.  But since that time Hierarchies are their own
thing and without the application data focus from the generic group
concept this looks much more like a system facilitation component.

So lets own that change and refactor back to component/system and
MscmpSyst*, etc.

This commit updates everything to do that with what exists now.  Unit
tests (that currently exist) are passing which is enough to give
confidence in the database and application server code refactoring.