Refactor MscmpSystMenu to MscmpSystNavigation
We're defining not just menus, but also command based navigation in
close proximity to the menus, so lets call this Component something more
inclusive of both concepts.  Naturally, later we'll have to deal with
naming and figuring out Components and component names to deal with the
actual user interface for both menus and command based navigation, but
that's a future me problem.