Add data table/API view commenting functions
Private database types and functions to support creating a standardized
table and column commenting format and database types and functions to
automatically API View documentation from their closely related data
tables.

This is code that almost certainly should have never been written... and
yet here we are.  The primary motivation was that a substantial amount
of documentation was being hand-copied from base database tables to
their related API Views.  At the start it was known that this was going
to amount to busy work which would be introduce maintenance related risk
to the API View documentation as documentation at the base table level
needed to be changed.  So the answer was not to hand copy the data table
documentation, but source the API View documentation from the data table
comments, including extras that might be needed.  Along the way,
standardize the bits of both the data table and API views so that we get
a higher consistency of documentation which presents better in the ERD
docs tool (and, yes, we also learned that the ERD tool will see markdown
in the comments and apply them accordingly).

While all that sounds great, the solution isn't great.  I do think it
will meet its goals in the end and will eventually save time and lead to
more accurate documentation.  But the costs are close enough to the
benefits that the smarter call would have been to just create a nice set
of IDE templates for the comment standards and be sure that we quality
check any base table updates for duplicating changes.  The downsides are
that it took a fair amount of time to come up with decent functions to
do the commenting and now developers have to learn our custom commenting
workflow to achieve our targeted consistency, rather than relying on the
standard database commenting tools they'd likely already know.  The
conceits of a vanity project.

Having said all that, to this system's credit it does work, produces
nice documentation, it does reduce redundancy and copy/paste work, and
will allow changes to formatting or the standard without having to touch
every database table or API view which would be considered major
otherwise.  So worth keeping for sure... just not worth starting given
hindsight.