Refactor for UUIDv7 (development configuration)
This commit introduces UUIDv7 for development purposes.  Currently this
requires a third party extension which won't be acceptable for a
releasable product, but is fine for development.  PostgreSQL 17 is
expected to include native support for UUIDv7; further we expect the
UUID generation API to be the same between the extension we're using and
the native support.  Since PostgreSQL 17 will definitely release before
we do, we can accept this arrangement.