Refactor user_* & syst_* comments_config_apiview
This changes names which started with user_record_* to user_* and
syst_record_* to syst_*; an example is `user_record_insert` is now just
`user_insert`, just like the name in comments_config_apiview_column.

The reason for this change is just to reduce the amount of stuff that's
needed to remember when creating documentation.  The previous names
were clearer and more explicit, but required remembering different names
for similar concepts across contexts.  The context should be enough to
know whether or not you're working at the view level or column level.