Add PowerShell based build scripts
This is a long outstanding task which was becoming more important to
finish.  This is especially true now that we're going to generate
database documentation for each Component individually.

The new script supports common cross-project actions such as updating
dependencies, building documentation, and similar actions which are
Component-by-Component and boilerplate-ish in nature.

Originally I started writing these scripts as Bash shells; our current
development environment is expected to be Linux based and it would be
the obvious first choice.  Having never written a real Bash script
before, I started following this conventional logic: the scripting
process itself convinced me otherwise.  In the end I landed on
PowerShell.  I'd never written a PowerShell script before either; but
the process was significantly more sane than what I found with Bash.

The reason PowerShell won over alternatives like my beloved Fish Shell
or something like Python or even Elixir scripts boils down to a few
basic facts.  First, while development is expected to be primarily Linux
driven, it's not intended to be exclusively Linux and there are cases
where building on Windows is necessary (release builds) so whatever we
scripting method we choose should ideally also be well supported on
Windows.  Python and Elixir both have runtime environments that may
not be present or may introduce runtime environment management issues
that are needless complicated for simple scripts.  PowerShell is well
supported on both Windows and Linux.  PowerShell, being a Microsoft
product, should be acceptable for use in enterprises and such where tool
selections could be controversial (yes, Elixir already has us off that
trail, but lets not push every front).  It will be more likely that a
Linux user will be more familiar with one-off installation of a tool
like PowerShell than a Windows user trying to similarly go off well
trodden paths.  Finally, PowerShell is a decent shell for this sort of
thing; the biggest issue is that it's such a rethink of the paradigm
that it's not substantially similar to anything that came before it...
especially the role and naming of Cmdlets.

These scripts are not examples of how best to write PowerShell; they
work and attempt to organize the code, but I expect there are plenty
bad and non-compliant with PowerShell standards and practices.