Add MscmpSystDb Mix tasks & DevSupport
Over time it has become apparent that we need a way to load the database
with schema and data outside the context of a running Component or
Subsystem.  This may facilitate documentation or development oriented
tools, like DataGrip, which want to inspect the database for symbol
references and such (instead of relying on the source code).  Alongside
this needs, its clear that we're duplicating a lot of code related to
supporting the similar need within running Components; this is the
DevSupport that's currently in all components and also similar Datastore
setup code in the testing setup.  Finally, we want a way that to load
the database with arbitrary database migrations so that our builds,
tests, and documentation generation can be more automated.

To this end, the following significant changes are included in this
commit:

  - Add a Mix task `loaddb` to be able to use Mix to create a basic,
    development supporting Datastore using default or supplied
    DatastoreOption values.

  - Add a Mix task `dropdb` to drop Datastores previously created with
    `loaddb`.

  - Add MscmpSystDb.Impl.DevSupport including functions to generate a
    simple development supporting DatastoreOptions object, with the
    ability to default most attributes to sensible values given the
    scenario.  This also includes the implementation details for both
    loading and dropping the Datastore used in the Mix tasks.

  - Expose the DevSupport module functions for generating
    DatastoreOptions, loading, and dropping the Datastore in the Public
    API of MscmpSystDb.  This way we can reduce the boilerplate code in
    each Component's DevSupport and testing infrastructure.

  - Update MscmpSystDb documentation so that the new DevSupport and Mix
    tasks are present.