Add MscmpSystSettings dev/test Support
Similar to recent changes in MscmpSystDb, this commit adds common
interactive development and testing support functions to the Component.
This should reduce boilerplate and allow for build scripts and other
automation to be more reliable and require less maintenance over time.

This commit includes a related update to the Elixir documentation.