Fix incorrect common table column comments
The general usage section of the table column comment generators always
left a warning that a column was not user maintainable and that is
simply not true in many cases; in some cases, such as user_description,
the reality is the opposite: the column is explicitly user updatable
under all conditions.

This commit allows for those exception cases.