Refactor SchemaSpy schemas to be more dynamic
Prior to this commit, the SchemaSpy calling portion of MsbmsBuildDocs
had a hard-coded list of all possible schemas that could exist.  This
resulted in documentation generation for schemas that didn't exist for
some components, which in turn made the docs just that much more
noisier.

Here we change the command line options to introspect the loaded
schemas and exclude anything that doesn't start with `ms_` (all of our
schemas should).