Refactor MssubMcp for DevSupport; DB Docs
This is not a full refactoring as MssubMcp does some special stuff
related to starting services using its own API; while these aren't in
formal tests, they constitute tests as nothing will pass if it can't
even start its own services in the prescribed way.

Net of it is that there's not as many changes here as in the other spots
as getting this less redundant will take some time once we're really
more focused on MssubMcp testing.