Add a DataGrip supporting database load
DataGrip works best when it has an introspected schema to work against.

This commit adds a DataGrip specific database load based on the MssubMcp
DB docs buildplan.

In addition, we'll add MscmpSystHierarchy and MscmpSystNav to the be
MssubMcp build.  They're not really ready for prime-time, but we're pre-
release and it's probably better to have these committed than not at
this stage.