Refactor initialize_enum into a procedure
PostgreSQL now supports `CREATE PROCEDURE` which is similar to the prior
function/`RETURNS void` function, but has some functional differences.

On working with the new procedures a bit, the older void returning
function most definitely still has a place if you don't need mid-
procedure commits; in fact because PostgreSQL procedures don't work in
DML and the void returning functions do work in DML, it may be that most
void returning functions should just stay that way.

In this case, initializing enumerations really matches that procedural
kind of definition is the kind of thing where you might want sub-
transaction like support.  So we convert that function.

I tried converting other void returning functions to procedures, but
almost immediately found cases where I wanted to work on the results of
a query: in function form, I can just include the function call in the
query, but with a procedure I'd have to loop through the results or
re-write the procedure to encapsulate the lookup query.

So for now, we'll err on writing void returning functions unless there's
a clear, unambiguous case to be made that the feature really is a
procedure.

This commit includes updating the calling semantics where the prior
function was called.