Fix generate_comments_function general_usage bug
The `general_usage` value of the comments_config object may come in as
an empty string; this is a very real possibility since the IDE templates
include this value as an empty string and requires the user to remember
to remove it.

Let's make the function forgiving of extraneous, but empty,
`general_usage` values.